requirements.txt
g4f
openai-whisper
numpy
elevenlabs
python-dotenv
moviepy
opencv-python
pydub